src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile src/requirements.in
#
amqp==5.1.1
    # via kombu
apache-beam[dataframe]==2.50.0
    # via -r src/requirements.in
asgiref==3.7.2
    # via django
async-timeout==4.0.3
    # via redis
backoff==2.2.1
    # via kinto-http
billiard==4.1.0
    # via celery
build==1.0.3
    # via pip-tools
celery[redis]==5.3.4
    # via -r src/requirements.in
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cloudpickle==2.2.1
    # via apache-beam
crcmod==1.7
    # via apache-beam
dill==0.3.1.1
    # via apache-beam
django==4.2.5
    # via
    #   -r src/requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
django-cors-headers==4.2.0
    # via -r src/requirements.in
django-filter==23.3
    # via -r src/requirements.in
djangorestframework==3.14.0
    # via -r src/requirements.in
dnspython==2.4.2
    # via pymongo
docopt==0.6.2
    # via hdfs
fastavro==1.8.3
    # via apache-beam
fasteners==0.19
    # via apache-beam
flink==1.0
    # via -r src/requirements.in
grpcio==1.59.0
    # via apache-beam
hdfs==2.7.2
    # via apache-beam
httplib2==0.22.0
    # via apache-beam
idna==3.4
    # via requests
kinto-http==11.0.1
    # via -r src/requirements.in
kombu==5.3.2
    # via celery
markdown==3.4.4
    # via -r src/requirements.in
numpy==1.24.4
    # via
    #   apache-beam
    #   pandas
    #   pyarrow
objsize==0.6.1
    # via apache-beam
orjson==3.9.7
    # via apache-beam
packaging==23.1
    # via build
pandas==1.5.3
    # via apache-beam
pip-tools==7.3.0
    # via -r src/requirements.in
prompt-toolkit==3.0.39
    # via click-repl
proto-plus==1.22.3
    # via apache-beam
protobuf==4.23.4
    # via
    #   apache-beam
    #   proto-plus
psycopg2-binary==2.9.8
    # via -r src/requirements.in
pyarrow==11.0.0
    # via apache-beam
pydot==1.4.2
    # via apache-beam
pymongo==4.5.0
    # via apache-beam
pyparsing==3.1.1
    # via
    #   httplib2
    #   pydot
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via
    #   apache-beam
    #   celery
    #   pandas
python-dotenv==1.0.0
    # via -r src/requirements.in
pytz==2023.3.post1
    # via
    #   apache-beam
    #   djangorestframework
    #   pandas
redis==4.6.0
    # via celery
regex==2023.8.8
    # via apache-beam
requests==2.31.0
    # via
    #   apache-beam
    #   hdfs
    #   kinto-http
six==1.16.0
    # via
    #   hdfs
    #   python-dateutil
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.8.0
    # via
    #   apache-beam
    #   asgiref
tzdata==2023.3
    # via celery
unidecode==1.3.7
    # via kinto-http
urllib3==2.0.5
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.8
    # via prompt-toolkit
wheel==0.41.2
    # via pip-tools
zstandard==0.21.0
    # via apache-beam

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
